+++ src/sharp-metadata.schema.ts
@@ -1,7 +1,7 @@
 // vim: tabstop=8 softtabstop=0 noexpandtab shiftwidth=8 nosmarttab
 // Derived from Sharp.Metadata in sharp.d.ts.
 
-import { z } from 'zod';
+import { z } from 'zod/v4';
 
 export const FormatEnum = z.enum([
 	'avif',