+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## v8.1.19
+- Bump to `zod@next` with `zod/v4` imports.
+
 ## v8.1.18
 - Add new metadata type `texture`, for GPU textures that are not supported by _Sharp_ or _FFmpeg_.
 - Rename to `@dsbunny/metadata-schema` for consistency among schemas.